Refactor graph components to common package
