+++ checkov/common/checks_infra/checks_parser.py
@@ -4,7 +4,7 @@ from checkov.common.graph.checks_infra.base_check import BaseGraphCheck
 from checkov.common.graph.checks_infra.base_parser import BaseGraphCheckParser
 from checkov.common.graph.checks_infra.enums import SolverType
 from checkov.common.graph.checks_infra.solvers.base_solver import BaseSolver
-from checkov.terraform.checks_infra.solvers import *
+from checkov.common.checks_infra.solvers import *
 
 operators_to_attributes_solver_classes = {
     "equals": EqualsAttributeSolver,
