+++ checkov/common/checks_infra/resources_types.py
