+++ checkov/common/checks_infra/solvers/__init__.py
@@ -1,0 +1,4 @@
+from checkov.common.checks_infra.solvers.attribute_solvers import *
+from checkov.common.checks_infra.solvers.complex_solvers import *
+from checkov.common.checks_infra.solvers.connections_solvers import *
+from checkov.common.checks_infra.solvers.filter_solvers import *
