+++ checkov/common/checks_infra/solvers/attribute_solvers/base_attribute_solver.py
