+++ checkov/common/checks_infra/solvers/attribute_solvers/not_contains_attribute_solver.py
