+++ checkov/common/checks_infra/solvers/attribute_solvers/not_ending_with_attribute_solver.py
