+++ checkov/common/checks_infra/solvers/attribute_solvers/not_starting_with_attribute_solver.py
