+++ checkov/common/checks_infra/solvers/attribute_solvers/starting_with_attribute_solver.py
@@ -1,14 +1,15 @@
 from typing import List, Optional, Any, Dict
 
 from checkov.common.graph.checks_infra.enums import Operators
-from checkov.terraform.checks_infra.solvers.attribute_solvers.base_attribute_solver import BaseAttributeSolver
+from checkov.common.checks_infra.solvers.attribute_solvers.base_attribute_solver import BaseAttributeSolver
 
 
-class ExistsAttributeSolver(BaseAttributeSolver):
-    operator = Operators.EXISTS
+class StartingWithAttributeSolver(BaseAttributeSolver):
+    operator = Operators.STARTING_WITH
 
     def __init__(self, resource_types: List[str], attribute: Optional[str], value: Any) -> None:
         super().__init__(resource_types=resource_types, attribute=attribute, value=value)
 
     def _get_operation(self, vertex: Dict[str, Any], attribute: Optional[str]) -> bool:
-        return vertex.get(attribute) is not None
+        attr = vertex.get(attribute)
+        return isinstance(attr, str) and attr.startswith(self.value)
