+++ checkov/common/checks_infra/solvers/complex_solvers/__init__.py
@@ -1,0 +1,2 @@
+from checkov.common.checks_infra.solvers.complex_solvers.or_solver import OrSolver
+from checkov.common.checks_infra.solvers.complex_solvers.and_solver import AndSolver
