+++ checkov/common/checks_infra/solvers/complex_solvers/and_solver.py
@@ -2,7 +2,7 @@ from typing import List, Any, Dict
 
 from checkov.common.graph.checks_infra.enums import Operators
 from checkov.common.graph.checks_infra.solvers.base_solver import BaseSolver
-from checkov.terraform.checks_infra.solvers.complex_solvers.base_complex_solver import BaseComplexSolver
+from checkov.common.checks_infra.solvers.complex_solvers.base_complex_solver import BaseComplexSolver
 from functools import reduce
 from operator import and_
