+++ checkov/common/checks_infra/solvers/complex_solvers/base_complex_solver.py
