+++ checkov/common/checks_infra/solvers/connections_solvers/base_connection_solver.py
