+++ checkov/common/checks_infra/solvers/connections_solvers/connection_exists_solver.py
@@ -2,7 +2,7 @@ import itertools
 from typing import List, Optional, Dict, Any, Tuple
 
 from checkov.common.graph.checks_infra.enums import Operators
-from checkov.terraform.checks_infra.solvers.connections_solvers.base_connection_solver import BaseConnectionSolver
+from checkov.common.checks_infra.solvers.connections_solvers.base_connection_solver import BaseConnectionSolver
 from networkx import edge_dfs, DiGraph
 from checkov.common.graph.graph_builder import CustomAttributes
 from checkov.terraform.graph_builder.graph_components.block_types import BlockType
