+++ checkov/common/checks_infra/solvers/filter_solvers/__init__.py
@@ -1,0 +1,1 @@
+from checkov.common.checks_infra.solvers.filter_solvers.within_filter_solver import WithinFilterSolver
