+++ checkov/common/checks_infra/solvers/filter_solvers/base_filter_solver.py
