--- checkov/terraform/checks_infra/solvers/__init__.py
@@ -1,4 +1,0 @@
-from checkov.terraform.checks_infra.solvers.attribute_solvers import *
-from checkov.terraform.checks_infra.solvers.complex_solvers import *
-from checkov.terraform.checks_infra.solvers.connections_solvers import *
-from checkov.terraform.checks_infra.solvers.filter_solvers import *
