--- checkov/terraform/checks_infra/solvers/attribute_solvers/__init__.py
@@ -1,12 +1,0 @@
-from checkov.terraform.checks_infra.solvers.attribute_solvers.any_attribute_solver import AnyResourceSolver
-from checkov.terraform.checks_infra.solvers.attribute_solvers.contains_attribute_solver import ContainsAttributeSolver
-from checkov.terraform.checks_infra.solvers.attribute_solvers.not_contains_attribute_solver import NotContainsAttributeSolver
-from checkov.terraform.checks_infra.solvers.attribute_solvers.ending_with_attribute_solver import EndingWithAttributeSolver
-from checkov.terraform.checks_infra.solvers.attribute_solvers.equals_attribute_solver import EqualsAttributeSolver
-from checkov.terraform.checks_infra.solvers.attribute_solvers.exists_attribute_solver import ExistsAttributeSolver
-from checkov.terraform.checks_infra.solvers.attribute_solvers.not_ending_with_attribute_solver import NotEndingWithAttributeSolver
-from checkov.terraform.checks_infra.solvers.attribute_solvers.not_equals_attribute_solver import NotEqualsAttributeSolver
-from checkov.terraform.checks_infra.solvers.attribute_solvers.not_exists_attribute_solver import NotExistsAttributeSolver
-from checkov.terraform.checks_infra.solvers.attribute_solvers.not_starting_with_attribute_solver import NotStartingWithAttributeSolver
-from checkov.terraform.checks_infra.solvers.attribute_solvers.starting_with_attribute_solver import StartingWithAttributeSolver
-from checkov.terraform.checks_infra.solvers.attribute_solvers.within_attribute_solver import WithinAttributeSolver
