--- checkov/terraform/checks_infra/solvers/complex_solvers/__init__.py
@@ -1,2 +1,0 @@
-from checkov.terraform.checks_infra.solvers.complex_solvers.or_solver import OrSolver
-from checkov.terraform.checks_infra.solvers.complex_solvers.and_solver import AndSolver
