--- checkov/terraform/checks_infra/solvers/connections_solvers/__init__.py
@@ -1,5 +1,0 @@
-from checkov.terraform.checks_infra.solvers.connections_solvers.and_connection_solver import AndConnectionSolver
-from checkov.terraform.checks_infra.solvers.connections_solvers.complex_connection_solver import ComplexConnectionSolver
-from checkov.terraform.checks_infra.solvers.connections_solvers.connection_exists_solver import ConnectionExistsSolver
-from checkov.terraform.checks_infra.solvers.connections_solvers.connection_not_exists_solver import ConnectionNotExistsSolver
-from checkov.terraform.checks_infra.solvers.connections_solvers.or_connection_solver import OrConnectionSolver
