--- checkov/terraform/checks_infra/solvers/filter_solvers/__init__.py
@@ -1,1 +1,0 @@
-from checkov.terraform.checks_infra.solvers.filter_solvers.within_filter_solver import WithinFilterSolver
